Fixed a bug in the query parser related to using quoted range queries.
The code was trying to use _tcscpy to strip off the first quote by
copying from the first character.  However, _tcscpy calls wcscpy which
clearly states in the manpage that the strings may not overlap.

I created a new function that uses memmove (which does allow memory
overlaps) to implement the quote stripping.